added layout markup, added header, refactored structure
