+++ src/components/App/App.jsx
@@ -1,10 +1,13 @@
 import React from 'react';
-import AuthPage from '../AuthPage/AuthPage';
+
+import AuthPage from '../AuthPage';
+import Layout from '../Layout';
 
 function App() {
   return (
     <>
-      <AuthPage />
+      {/* <AuthPage /> */}
+      <Layout />
     </>
   );
 }
