+++ src/components/Layout.jsx
@@ -1,0 +1,13 @@
+import React, { Component } from 'react';
+
+import Header from './Header';
+
+class Layout extends Component {
+  render() {
+    return (
+      <Header />
+    );
+  }
+}
+
+export default Layout;
