+++ src/components/MapPage.jsx
@@ -1,0 +1,9 @@
+import React, { Component } from 'react';
+
+class MapPage extends Component {
+  render() {
+    return <div>Map Page</div>;
+  }
+}
+
+export default MapPage;
