+++ src/components/ProfilePage.jsx
@@ -1,0 +1,9 @@
+import React, { Component } from 'react';
+
+class ProfilePage extends Component {
+  render() {
+    return <div>Profile Page</div>;
+  }
+}
+
+export default ProfilePage;
